+++ spec/spec_helper.rb
@@ -3,3 +3,5 @@ require 'rspec'
 
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 require 'socky-client'
+
+Socky.logger.level = Logger::ERROR